Cleaned up the scraper to fix the site.
It appears as though the blended learning Wordpress install has started returning blank pages occasionally. I believe these are likely cache misses. So as a workaround, I just prime the cache on the remote server before running the scrape.

Also, a bad thing happened: the scrape overwrote the previous database without sanity checking that it had good data to replace it with. This turns a failure to update into a catastrophic failure. So now I check to see if there are results before wiping the DB.